Move coroutine cleanup regressions into TestJob
